sample-services/requirements.txt
fastapi==0.115.8
httpx==0.28.1
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-sdk==1.30.0
prometheus-client==0.21.1
uvicorn[standard]==0.34.0
